lesson 11 small refactoring
