+++ NotePad/app/src/main/java/com/example/mynotepad/interfaces/IDrawerHeaderHandler.java
@@ -1,4 +1,4 @@
-package com.example.mynotepad.intefaces;
+package com.example.mynotepad.interfaces;
 
 public interface IDrawerHeaderHandler {
